+++ scrapy_cffi/hooks/pipelines.py
@@ -1,4 +1,4 @@
-from typing import Protocol, Union, TYPE_CHECKING, Dict
+from typing import Protocol, TYPE_CHECKING, Dict
 if TYPE_CHECKING:
     from ..models.api import SingalInfo
 
@@ -8,7 +8,7 @@ class SignalHooks(Protocol):
 class SessionHooks(Protocol):
     def mark_end(self, session_id: str) -> None: ...
 
-    async def session_end_cookies(self, session_id: str) -> Union[Dict, None]: ...
+    def get_session_cookies(self, session_id: str) -> Dict: ...
 
 class _PipelinesHooks(Protocol):
     session: SessionHooks