+++ scrapy_cffi/utils/__init__.py
@@ -1,3 +1,4 @@
+from .email import *
 from .fd import FDUtil
 from .algorithm import *
 from .jsonLoad import *