+++ CHANGELOG.md
@@ -8,6 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.1.6] - 2025-09-07
+### Added
+- Spider hooks get_session_cookies
+- Email send util
+- StatisticsExtension
+
+## Fixed
+- Request Interceptor behavior bug (return a Response)
+
+---
+
 ## [0.1.5] - 2025-09-06
 ### Added
 - Initial PyPI release